runtime_table + JIT: widen TableHash returns to int64_t; fix JIT vec3 stride width
Two distinct correctness bugs surfaced after the Phase 2 widening, both fixed here.

TableHash<KT> int narrowing (Copilot + Boris): find/reserve/erase/insertNew
returned int while internally indexing uint64 capacity. For tables > INT_MAX,
the (int) cast at every return wrapped silently; reserveInternal's rehash
loop then did pHashes[wild_negative] / memcpy(data + wild*size) — UB. Widened
the four methods + the rehash-loop local + all C++ callers (aot.h, runtime_
table_nodes.h, simulate_fusion_tableindex.cpp, json_scan.cpp, daScriptC.cpp)
to int64_t. The -1 not-found sentinel still fits.

JIT helpers (module_jit.cpp jit_table_at/find and module_builtin_rtti.cpp
tableFindValue) keep their int32_t return because their LLVM-side codegen
expects i32; added explicit (int32_t) cast + TODO Phase 7 marker to make the
narrowing visible.

JIT verifier failure (CI 64-bit jobs across the board): the for-loop "first"
pseudoiterator over arrays now produces an i64 `tail` (Array.size is i64),
which flowed into build_array_index. Its vec3 special-case multiplied
tidx × ConstI32(12) — i64 × i32 — failing LLVMVerifyFunction with
"Both operands to a binary operator are not of the same type! %5 = mul i64
%3, i32 12". Build the stride constant at LLVMTypeOf(tidx)'s width instead.

skills/make_pr.md: add a JIT smoke step + quick-ref row. The two suggested
greps catch the verifier-error class of regressions locally even when
Windows clang-cl can't link the JIT'd DLL.

+++ include/daScript/simulate/aot.h
@@ -1139,21 +1139,21 @@ namespace das {
         __forceinline TV & operator () ( const TK & key, Context * __context__ ) {
             TableHash<TK> thh(__context__,sizeof(TV));
             auto hfn = hash_function(*__context__, key);
-            int index = thh.reserve(*this, key, hfn);
+            int64_t index = thh.reserve(*this, key, hfn);
             return ((TV *)data)[index];
         }
         static __forceinline TV * safe_index ( THIS_TYPE * that, const TK & key, Context * __context__ ) {
             if (!that) return nullptr;
             TableHash<TK> thh(__context__,sizeof(TV));
             auto hfn = hash_function(*__context__, key);
-            int index = thh.find(*that, key, hfn);
+            int64_t index = thh.find(*that, key, hfn);
             return index==-1 ? nullptr : ((TV *)that->data) + index;
         }
         static __forceinline const TV * safe_index ( const THIS_TYPE * that, const TK & key, Context * __context__ ) {
             if (!that) return nullptr;
             TableHash<TK> thh(__context__,sizeof(TV));
             auto hfn = hash_function(*__context__, key);
-            int index = thh.find(*that, key, hfn);
+            int64_t index = thh.find(*that, key, hfn);
             return index==-1 ? nullptr : ((const TV *)that->data) + index;
         }
     };
@@ -2911,7 +2911,7 @@ namespace das {
         TK key = (TK) _key;
         auto hfn = hash_function(*context, key);
         TableHash<TK> thh(context,sizeof(TV));
-        int index = thh.find(tab, key, hfn);
+        int64_t index = thh.find(tab, key, hfn);
         return (TV *) ( index!=-1 ? tab.data + index * sizeof(TV) : nullptr );
     }
